requirements.txt
flask
gunicorn
cryptography
flask-socketio
eventlet
Flask-Failsafe
mysql-connector-python
